Fix test for converters (#184)
* Fix converters for pipeline.coverage

* Fix converter

* Support negative numbers

* review

+++ NGitLab/Impl/Json/DoubleConverter.cs
@@ -10,12 +10,12 @@ namespace NGitLab.Impl.Json
         public override double Read(ref Utf8JsonReader reader, Type type, JsonSerializerOptions options)
         {
             if (reader.TokenType == JsonTokenType.Null)
-                return 0.0;
+                return 0.0d;
 
             if (reader.TokenType == JsonTokenType.String)
             {
                 if (double.TryParse(reader.GetString(), NumberStyles.Any, CultureInfo.InvariantCulture, out var d))
-                    return (int)d;
+                    return d;
             }
 
             return reader.GetDouble();
